Create README.md
I have paid for a service but it keep getting cut off. I am hoping to get as much help as possible. I would like to learn how to get codes. How to use them.where to find them or how to make them. Were to get premiums codes to.